requirements.txt
numpy==1.20.3
torch==1.8.2
torchvision==0.9.2
tqdm==4.62.2
Pillow==8.3.1
scipy==1.6.2
opencv-python==4.5.5.64
